+++ tomcat/src/main/java/nextstep/jwp/controller/RegisterController.java
@@ -1,7 +1,7 @@
 package nextstep.jwp.controller;
 
 import nextstep.jwp.model.User;
-import nextstep.jwp.model.UserService;
+import nextstep.jwp.application.UserService;
 import org.apache.coyote.http11.request.HttpRequest;
 import org.apache.coyote.http11.response.HttpResponse;
